Prevent overwrites of development data when Supervisor recreates containers

+++ development/start.sh
@@ -2,13 +2,23 @@
 destination=/app/
 source=dist/
 
+destination_timestamp=$destination.build_timestamp
+source_timestamp=$source.build_timestamp
+
+start_time=$(date +%s)
+
 # Import env variables to container
 . .env_vars
 
 echo "Running in hot reload mode. Setting up the development environment..."
-start_time=$(date +%s)
 
-rsync --archive --delete --inplace $source $destination
+if [ -e "$destination_timestamp" ] && [ "$(cat ${destination_timestamp})" = "$(cat ${source_timestamp})" ]; then
+    echo "App is up to date."
+else
+    # Update volume with latest UI files
+    rsync --archive --delete --inplace $source $destination
+    echo "App updated."
+fi
 
 echo "Development environment ready. Setup took $(($(date +%s)-$start_time)) seconds."
